fix: baby weight, height 타입 변경

+++ app/src/main/java/com/example/baby/screen/BabyRegisterScreen.kt
@@ -476,8 +476,8 @@ fun BabyRegisterButton(
                     Baby(
                         name = name,
                         gender = gender,
-                        birthWeight = weight.toInt(),
-                        birthHeight = height.toInt(),
+                        birthWeight = weight,
+                        birthHeight = height,
                         imageUrl = "",
                         dateTime = LocalDateTime.of(
                             LocalDate.of(
